+++ ceammc/ext/src/flow/seq_nbangs.h
@@ -23,6 +23,8 @@ class SeqNBangs : public BaseObject {
     IntProperty* n_ { 0 };
     SeqTimeGrain* interval_ { 0 };
     IntProperty* beat_division_ { 0 };
+    FloatProperty* accel_ { 0 };
+    FloatProperty* accel_curve_ { 0 };
     size_t counter_;
     ClockLambdaFunction clock_;
 
@@ -41,6 +43,9 @@ public:
 
     void resetCycleCounter() { }
     void resetSequenceCounter() { counter_ = 0; }
+
+private:
+    t_float nextBang() const;
 };
 
 using SeqNBangsT = SequencerIFace<SeqNBangs>;